make GtSequenceBuffers compile on 32bit systems

+++ src/core/sequence_buffer_embl.c
@@ -91,7 +91,7 @@ parse_next_line(GtSequenceBuffer *sb, GtEMBLParserLineCode *lc,
   (*currentfileread)++;
   if (currentchar == NEWLINESYMBOL) {
     gt_error_set(err, "2-character line code not found in line %lu",
-                 pvt->linenum - 1);
+                 (unsigned long) pvt->linenum - 1);
     return -2;
   }
   linecode[1] = currentchar;
@@ -118,7 +118,7 @@ parse_next_line(GtSequenceBuffer *sb, GtEMBLParserLineCode *lc,
       if (!isspace(currentchar)) {
         gt_error_set(err, "3 blanks expected between line code and content "
                           "in line %lu",
-                     pvt->linenum - 1);
+                     (unsigned long) pvt->linenum - 1);
         return -2;
       }
     }
@@ -253,7 +253,8 @@ static int gt_sequence_buffer_embl_advance(GtSequenceBuffer *sb, GtError *err)
         /* only a terminator may come after a sequence */
         if (lc != SEQUENCE) {
           gt_error_set(err, "unterminated sequence in line %lu of file %s",
-                            pvt->linenum, gt_str_array_get(pvt->filenametab,
+                            (unsigned long) pvt->linenum,
+                            gt_str_array_get(pvt->filenametab,
                                                 (unsigned long) pvt->filenum));
           return -1;
         }
